requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
anyio==4.12.1
asgiref==3.11.1
asttokens==3.0.1
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.3
cron-descriptor==1.4.5
cryptography==46.0.5
debugpy==1.8.20
decorator==5.2.1
dj-database-url==3.1.0
Django==6.0.2
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-redis==6.0.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
executing==2.2.1
filelock==3.20.0
fsspec==2025.12.0
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
gunicorn==25.1.0
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kafka-python==2.3.0
kombu==5.6.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.3
numpy==2.4.2
oauth2==1.9.0.post1
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.3.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
sympy==1.14.0
textblob==0.19.0
tokenizers==0.22.2
torch==2.10.0+cpu
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
